+++ src/n8n/nodes/10-read/command-parser__926eb875-5735-4746-a0a4-7801b8db586f.js
@@ -12,13 +12,7 @@
  */
 'use strict';
 
-const { mdv2Message } = (() => {
-  try {
-    return require('/data/src/libs/telegram-markdown.js');
-  } catch (err) {
-    return require('../../../libs/telegram-markdown.js');
-  }
-})();
+const { mdv2Message } = require('@igasovic/n8n-blocks/shared/telegram-markdown.js');
 
 module.exports = async function run(ctx) {
   const { $input, $json, $items, $node, $env, helpers } = ctx;